requirements.txt
streamlit
python-dotenv
requests
pytesseract
pillow
python-docx
pymupdf
openai
SpeechRecognition
pyaudio
streamlit-webrtc
groq